+++ src/test/java/cn/yuqi/mini/spring/beans/factory/BeanFactoryTest.java
@@ -3,6 +3,7 @@ package cn.yuqi.mini.spring.beans.factory;
 import cn.yuqi.mini.spring.beans.PropertyValue;
 import cn.yuqi.mini.spring.beans.PropertyValues;
 import cn.yuqi.mini.spring.beans.factory.config.BeanDefinition;
+import cn.yuqi.mini.spring.beans.factory.config.BeanReference;
 import cn.yuqi.mini.spring.beans.factory.support.CglibSubclassingInstantiationStrategy;
 import cn.yuqi.mini.spring.beans.factory.support.DefaultListableBeanFactory;
 import org.junit.jupiter.api.Test;
@@ -70,4 +71,34 @@ class BeanFactoryTest {
 //            return "hello";
 //        }
 //    }
+
+    @Test
+    public void testPopulateBeanWithBean() throws Exception {
+        DefaultListableBeanFactory beanFactory = new DefaultListableBeanFactory();
+
+        //注册Car实例
+        PropertyValues propertyValuesForCar = new PropertyValues();
+        propertyValuesForCar.addPropertyValue(new PropertyValue("brand", "porsche"));
+        BeanDefinition carBeanDefinition = new BeanDefinition(Car.class);
+        carBeanDefinition.setPropertyValues(propertyValuesForCar);
+        beanFactory.registerBeanDefinition("car", carBeanDefinition);
+
+        //注册Person实例
+        PropertyValues propertyValuesForPerson = new PropertyValues();
+        propertyValuesForPerson.addPropertyValue(new PropertyValue("name", "derek"));
+        propertyValuesForPerson.addPropertyValue(new PropertyValue("age", 18));
+        //Person实例依赖Car实例
+        propertyValuesForPerson.addPropertyValue(new PropertyValue("car", new BeanReference("car")));
+        BeanDefinition beanDefinition = new BeanDefinition(Person.class);
+        beanDefinition.setPropertyValues(propertyValuesForPerson);
+        beanFactory.registerBeanDefinition("person", beanDefinition);
+
+        Person person = (Person) beanFactory.getBean("person");
+        System.out.println(person);
+        assertThat(person.getName()).isEqualTo("derek");
+        assertThat(person.getAge()).isEqualTo(18);
+        Car car = person.getCar();
+        assertThat(car).isNotNull();
+        assertThat(car.getBrand()).isEqualTo("porsche");
+    }
 }